+++ packages/kitchensink/service-fastify-gql/service-fastify-gql.test.js
@@ -1,0 +1,6 @@
+describe('service-fastify-gql', () => {
+  it('should load a graphql query', async () => {
+    const res = await get('/');
+    expect(res.message).toBe('+ok');
+  });
+});