+++ packages/service-fastify-gql/src/hooks.js
@@ -1,6 +1,4 @@
-const { SERVICE } = require('@forrestjs/hooks');
-
-const SERVICE_NAME = `${SERVICE} fastify-gql`;
+const SERVICE_NAME = `fastify-gql`;
 const FASTIFY_GQL_EXTEND_SCHEMA = `${SERVICE_NAME}/extend-schema`;
 const FASTIFY_GQL_EXTEND_CONTEXT = `${SERVICE_NAME}/extend-context`;