+++ packages/service-fastify-gql/CHANGELOG.md
@@ -1,2 +1,6 @@
 # Change Log
 
+## v4.0.0
+
+- Remove Babel
+- Update to 3.0.0-lambda.0